+++ EZWorkout/Rec/RecordWorkoutMasterViewController+Commands.swift
@@ -1,0 +1,17 @@
+
+import UIKit
+
+extension RecordWorkoutMasterViewController{
+  
+  func Undo(){
+    
+    //TODO undo other actions?
+    if exercises != nil , exercises!.count > 0{
+      let exercise = exercises!.removeLast()
+      createFeedbackMessage(topMsg: "removing exercise:", bottomMsg: exercise.name)
+     centerController?.reloadTable()
+    }else{
+      createFeedbackMessage(topMsg: "nothing to undo", bottomMsg: "")
+    }
+  }
+}
