+++ EZWorkout/Rec/RecordWorkoutMasterViewController.swift
@@ -14,14 +14,14 @@ class RecordWorkoutMasterViewController: UIViewController {
   
   
   
-    var centerController: CenterViewController?
+  var centerController: CenterViewController?
   
   let speechInputManager = SpeechInputManager()
   var startDate: Date?
   
   var workoutModel: WorkoutModel?
   var exercises: [ExerciseModel]?
-  
+  var lastExercise: ExerciseModel?
   
   var isWorkingOut = false
   
@@ -146,49 +146,4 @@ class RecordWorkoutMasterViewController: UIViewController {
   
 }
 
-extension RecordWorkoutMasterViewController : RecognizerReturnDelegate {
-  
-  func recieveInProgress(speech: String) {
-    feebBackLabel.text = speech
-  }
-  
-  func recieve(speech: String) {
-    print("speech recieved: \(speech)")
-    
-    //if working out, check exercises first
-    
-      //process for exercise data
-      var exerciseModel = speechInputManager.findExercise(input: speech)
-      if let exerciseModel = exerciseModel {
-        print("found an exercise with the name \(exerciseModel.name)")
-        if isWorkingOut{
-          createExerciseView(exerciseModel: exerciseModel)
-        }else{
-          //TODO auto start workout
-          createFeedbackMessage(topMsg: "you must start a workout first", bottomMsg: "")
-        }
-        
-        return
-      }
-    
-    
-    //then check the commands
-    let cmd = speechInputManager.findCommand(inputSpeech: speech)
-    
-    if let cmd = cmd{
-      switch (cmd){
-      case .StartWorkout:
-        startWorkout()
-      case .EndWorkout:
-        endWorkout()
-      }
-    }else{
-      print("did not understand command")
-      createFeedbackMessage(topMsg: "did not understand command", bottomMsg: "\'\(speech)\'")
-    }
-    
-    
-  }
-  
-  
-}
+
