+++ EZWorkout/SpeechParsing/SpeechInputManager/SpeechInputManager+Commands.swift
@@ -9,16 +9,23 @@ extension SpeechInputManager{
     let cleanInput = sanitize(input: inputSpeech)
     print("clean input = \(cleanInput)")
     //TODO expand with more commands
+    
     if cleanInput.contains("startwork"){
       return SpeechCommand.StartWorkout
     }
+    
     if (cleanInput.contains("endwork") || cleanInput.contains("andwork") || cleanInput.contains("finishworkout")) {
       return SpeechCommand.EndWorkout
     }
+    
+    if (cleanInput.contains("undo")  ) {
+      return SpeechCommand.Undo
+    }
+    
     return nil
   }
 }
 
 enum SpeechCommand{
-  case StartWorkout,EndWorkout
+  case StartWorkout,EndWorkout,Undo
 }
